actions [nfc]: Move logOutAccount and unregisterToken under lib/model
plus moving and refactoring the tests to match
