+++ test/widgets/actions_test.dart
@@ -1,11 +1,9 @@
 import 'dart:convert';
 
 import 'package:checks/checks.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:http/http.dart' as http;
-import 'package:zulip/api/exception.dart';
 import 'package:zulip/api/model/initial_snapshot.dart';
 import 'package:zulip/api/model/model.dart';
 import 'package:zulip/api/model/narrow.dart';
@@ -13,14 +11,11 @@ import 'package:zulip/api/route/messages.dart';
 import 'package:zulip/model/localizations.dart';
 import 'package:zulip/model/narrow.dart';
 import 'package:zulip/model/store.dart';
-import 'package:zulip/notifications/receive.dart';
 import 'package:zulip/widgets/actions.dart';
 
 import '../api/fake_api.dart';
 import '../example_data.dart' as eg;
 import '../model/binding.dart';
-import '../model/store_checks.dart';
-import '../model/test_store.dart';
 import '../model/unreads_checks.dart';
 import '../stdlib_checks.dart';
 import 'dialog_checks.dart';
@@ -52,162 +47,6 @@ void main() {
     await tester.pump();
     context = tester.element(find.byType(Placeholder));
   }
-
-  /// Creates and caches a new [FakeApiConnection] in [TestGlobalStore].
-  ///
-  /// In live code, [unregisterToken] makes a new [ApiConnection] for the
-  /// unregister-token request instead of reusing the store's connection.
-  /// To enable callers to prepare responses for that request, this function
-  /// creates a new [FakeApiConnection] and caches it in [TestGlobalStore]
-  /// for [unregisterToken] to pick up.
-  ///
-  /// Call this instead of just turning on
-  /// [TestGlobalStore.useCachedApiConnections] so that [unregisterToken]
-  /// doesn't try to call `close` twice on the same connection instance,
-  /// which isn't allowed. (Once by the unregister-token code
-  /// and once as part of removing the account.)
-  FakeApiConnection separateConnection() {
-    testBinding.globalStore
-      ..clearCachedApiConnections()
-      ..useCachedApiConnections = true;
-    return testBinding.globalStore
-      .apiConnectionFromAccount(eg.selfAccount) as FakeApiConnection;
-  }
-
-  String unregisterApiPathForPlatform(TargetPlatform platform) {
-    return switch (platform) {
-      TargetPlatform.android => '/api/v1/users/me/android_gcm_reg_id',
-      TargetPlatform.iOS     => '/api/v1/users/me/apns_device_token',
-      _                      => throw Error(),
-    };
-  }
-
-  void checkSingleUnregisterRequest(
-    FakeApiConnection connection, {
-    String? expectedToken,
-  }) {
-    final subject = check(connection.takeRequests()).single.isA<http.Request>()
-      ..method.equals('DELETE')
-      ..url.path.equals(unregisterApiPathForPlatform(defaultTargetPlatform));
-    if (expectedToken != null) {
-      subject.bodyFields.deepEquals({'token': expectedToken});
-    }
-  }
-
-  group('logOutAccount', () {
-    testWidgets('smoke', (tester) async {
-      await prepare(tester, skipAssertAccountExists: true);
-      check(testBinding.globalStore).accountIds.single.equals(eg.selfAccount.id);
-      const unregisterDelay = Duration(seconds: 5);
-      assert(unregisterDelay > TestGlobalStore.removeAccountDuration);
-      final newConnection = separateConnection()
-        ..prepare(delay: unregisterDelay, json: {'msg': '', 'result': 'success'});
-
-      final future = logOutAccount(testBinding.globalStore, eg.selfAccount.id);
-      // Unregister-token request and account removal dispatched together
-      checkSingleUnregisterRequest(newConnection);
-      check(testBinding.globalStore.takeDoRemoveAccountCalls())
-        .single.equals(eg.selfAccount.id);
-
-      await tester.pump(TestGlobalStore.removeAccountDuration);
-      await future;
-      // Account removal not blocked on unregister-token response
-      check(testBinding.globalStore).accountIds.isEmpty();
-      check(connection.isOpen).isFalse();
-      check(newConnection.isOpen).isTrue(); // still busy with unregister-token
-
-      await tester.pump(unregisterDelay - TestGlobalStore.removeAccountDuration);
-      check(newConnection.isOpen).isFalse();
-    });
-
-    testWidgets('unregister request has an error', (tester) async {
-      await prepare(tester, skipAssertAccountExists: true);
-      check(testBinding.globalStore).accountIds.single.equals(eg.selfAccount.id);
-      const unregisterDelay = Duration(seconds: 5);
-      assert(unregisterDelay > TestGlobalStore.removeAccountDuration);
-      final exception = ZulipApiException(
-        httpStatus: 401,
-        code: 'UNAUTHORIZED',
-        data: {"result": "error", "msg": "Invalid API key", "code": "UNAUTHORIZED"},
-        routeName: 'removeEtcEtcToken',
-        message: 'Invalid API key',
-      );
-      final newConnection = separateConnection()
-        ..prepare(delay: unregisterDelay, exception: exception);
-
-      final future = logOutAccount(testBinding.globalStore, eg.selfAccount.id);
-      // Unregister-token request and account removal dispatched together
-      checkSingleUnregisterRequest(newConnection);
-      check(testBinding.globalStore.takeDoRemoveAccountCalls())
-        .single.equals(eg.selfAccount.id);
-
-      await tester.pump(TestGlobalStore.removeAccountDuration);
-      await future;
-      // Account removal not blocked on unregister-token response
-      check(testBinding.globalStore).accountIds.isEmpty();
-      check(connection.isOpen).isFalse();
-      check(newConnection.isOpen).isTrue(); // for the unregister-token request
-
-      await tester.pump(unregisterDelay - TestGlobalStore.removeAccountDuration);
-      check(newConnection.isOpen).isFalse();
-    });
-  });
-
-  group('unregisterToken', () {
-    testWidgets('smoke, happy path', (tester) async {
-      await prepare(tester, ackedPushToken: '123');
-
-      final newConnection = separateConnection()
-        ..prepare(json: {'msg': '', 'result': 'success'});
-      final future = unregisterToken(testBinding.globalStore, eg.selfAccount.id);
-      await tester.pump(Duration.zero);
-      await future;
-      checkSingleUnregisterRequest(newConnection, expectedToken: '123');
-      check(newConnection.isOpen).isFalse();
-    }, variant: const TargetPlatformVariant({TargetPlatform.android, TargetPlatform.iOS}));
-
-    testWidgets('fallback to current token if acked is missing', (tester) async {
-      await prepare(tester, ackedPushToken: null);
-      NotificationService.instance.token = ValueNotifier('asdf');
-
-      final newConnection = separateConnection()
-        ..prepare(json: {'msg': '', 'result': 'success'});
-      final future = unregisterToken(testBinding.globalStore, eg.selfAccount.id);
-      await tester.pump(Duration.zero);
-      await future;
-      checkSingleUnregisterRequest(newConnection, expectedToken: 'asdf');
-      check(newConnection.isOpen).isFalse();
-    });
-
-    testWidgets('no error if acked token and current token both missing', (tester) async {
-      await prepare(tester, ackedPushToken: null);
-      NotificationService.instance.token = ValueNotifier(null);
-
-      final newConnection = separateConnection();
-      final future = unregisterToken(testBinding.globalStore, eg.selfAccount.id);
-      await tester.pumpAndSettle();
-      await future;
-      check(newConnection.takeRequests()).isEmpty();
-    });
-
-    testWidgets('connection closed if request errors', (tester) async {
-      await prepare(tester, ackedPushToken: '123');
-
-      final newConnection = separateConnection()
-        ..prepare(exception: ZulipApiException(
-            httpStatus: 401,
-            code: 'UNAUTHORIZED',
-            data: {"result": "error", "msg": "Invalid API key", "code": "UNAUTHORIZED"},
-            routeName: 'removeEtcEtcToken',
-            message: 'Invalid API key',
-          ));
-      final future = unregisterToken(testBinding.globalStore, eg.selfAccount.id);
-      await tester.pump(Duration.zero);
-      await future;
-      checkSingleUnregisterRequest(newConnection, expectedToken: '123');
-      check(newConnection.isOpen).isFalse();
-    });
-  });
 
   group('markNarrowAsRead', () {
     testWidgets('smoke test on modern server', (tester) async {
